mcp-http-echo-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-http-echo-server"
version = "1.0.0"
description = "A dual-mode (stateful/stateless) MCP echo server with 21 comprehensive debugging tools built on FastMCP 2.0"
authors = [
    { name = "Andreas Trawoeger", email = "atrawog@gmail.com" }
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "asyncio",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
]
redis = [
    "redis>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/atrawog/mcp-http-proxy/tree/main/mcp-http-echo-server"
Repository = "https://github.com/atrawog/mcp-http-proxy/tree/main/mcp-http-echo-server"
Issues = "https://github.com/atrawog/mcp-http-proxy/issues"
Documentation = "https://atrawog.github.io/mcp-http-proxy"

[project.scripts]
mcp-http-echo-server = "mcp_http_echo_server.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_http_echo_server"]